Now always return ε₁ = ∞ when we exceed T_cap
From my understanding of the pseudocode, this is actually the intended
behavior. Without this change, if verification proceeds past round i =
0, then ε₁ is generally some small value (e.g. 0.16), which invariably
leads to setting `t = t + 1`.

Technically, I don't think checking the finiteness of `eps1` when updating
`t` is necessary. However, the behavior of min/max w.r.t. infinity is
not obviously correct (although in the case of the Rust stdlib it
actually *is* correct), so I took this approach.

Also added logging of verification samples